+++ pkg/util/allowed_tenants.go
@@ -1,5 +1,7 @@
 package util
 
+const GlobalMarkersDir = "__markers__"
+
 // AllowedTenants that can answer whether tenant is allowed or not based on configuration.
 // Default value (nil) allows all tenants.
 type AllowedTenants struct {
@@ -34,6 +36,11 @@ func NewAllowedTenants(enabled []string, disabled []string) *AllowedTenants {
 }
 
 func (a *AllowedTenants) IsAllowed(tenantID string) bool {
+	if tenantID == GlobalMarkersDir {
+		// __markers__ is reserved for global markers and no tenant should be allowed to have that name.
+		return false
+	}
+
 	if a == nil {
 		return true
 	}